Added Building Controller.

+++ src/main/java/com/emse/spring/faircorp/dao/BuildingDaoCustom.java
@@ -6,5 +6,5 @@ import com.emse.spring.faircorp.model.Room;
 import java.util.List;
 
 public interface BuildingDaoCustom {
-    List<Light> findBuildingLights(Integer buildingId);
+    List<Light> findBuildingLights(Long buildingId);
 }
